Fix camera resolution

+++ src/Camera.cpp
@@ -16,7 +16,9 @@ void Camera::Matrix(float FOVdeg, float nearPlane, float farPlane, Shader& shade
     glm::mat4 projection = glm::mat4(1.0f);
 
     view = glm::lookAt(Position, Position + Orientation, Up);
-    projection = glm::perspective(glm::radians(FOVdeg), (float)(width / height), nearPlane, farPlane);
+
+    // Fix: Cast to float before division to get correct aspect ratio
+    projection = glm::perspective(glm::radians(FOVdeg), (float)width / (float)height, nearPlane, farPlane);
 
     glUniformMatrix4fv(glGetUniformLocation(shader.ID, uniform), 1, GL_FALSE, glm::value_ptr(projection * view));
 }
